Modify signature of apply in transform API [ci skip]
The apply method takes a vector of queries rather than a single query.
The returned values are returned as references.

This new API allows to apply a transformation only once to the
transition system regardless of the number of queries. Before, we
applied the transformation for each pair of transition system and
query.

Moreover, this new API allows a transformation to modify the state
type after a transition system or query is processed. Before, we only
allowed to create a new state type in the constructor of the derived
transformation class before a transition system or query was
processed. This did not allow, for instance, to update a state type
with new fresh variables added during the transformation of a
particular transition system or query.