Remove more array terms [ci skip]
Before we only removed array reads assuming that the array term was
another array read or an array variable. This commit also removes
array writes and removes array terms even if the array term of the
array write/read is a complicated ite term.

The removal of array writes is done by translating them to array reads
and it requires auxiliary variables. This is not ideal so future work
should remove array writes without adding these auxiliary variables.